rename tests package to rostest
