+++ tests/integration/rostest/test_01_cloud_config.py
@@ -1,5 +1,5 @@
 import pytest
-import rancherostest.util as u
+import rostest.util as u
 import string
 import subprocess
 import yaml
