+++ tests/integration/rostest/test_05_custom_docker_in_persistent_console.py
@@ -1,5 +1,5 @@
 import pytest
-import rancherostest.util as u
+import rostest.util as u
 import subprocess
