+++ tests/integration/setup.py
@@ -4,7 +4,7 @@ setup(
     name='RancherOSIntegrationTests',
     version='0.1',
     packages=[
-        'rancherostest'
+        'rostest'
     ],
     license='ASL 2.0',
     long_description=open('README.txt').read(),
