+++ tests/integration/tox.ini
@@ -4,11 +4,11 @@ skipsdist=True
 
 [testenv]
 deps=-rrequirements.txt
-commands=py.test -s --durations=20 rancherostest {posargs}
+commands=py.test -s --durations=20 rostest {posargs}
 
 [testenv:flake8]
 deps=flake8
-commands=flake8 rancherostest
+commands=flake8 rostest
 
 [flake8]
 max-line-length = 120
